move artifact to common
