+++ pkg/api/aim/services/run/service.go
@@ -16,9 +16,9 @@ import (
 	"github.com/G-Research/fasttrackml/pkg/api/aim/common"
 	"github.com/G-Research/fasttrackml/pkg/api/aim/dao/models"
 	"github.com/G-Research/fasttrackml/pkg/api/aim/dao/repositories"
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/services/artifact/storage"
 	"github.com/G-Research/fasttrackml/pkg/common/api"
 	"github.com/G-Research/fasttrackml/pkg/common/dao/types"
+	"github.com/G-Research/fasttrackml/pkg/common/services/artifact/storage"
 )
 
 // allowed batch actions.
