+++ pkg/api/mlflow/controller/artifacts.go
@@ -11,10 +11,10 @@ import (
 	"github.com/rotisserie/eris"
 	log "github.com/sirupsen/logrus"
 
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api/request"
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api/response"
 	"github.com/G-Research/fasttrackml/pkg/api/mlflow/common"
 	"github.com/G-Research/fasttrackml/pkg/common/api"
+	"github.com/G-Research/fasttrackml/pkg/common/api/request"
+	"github.com/G-Research/fasttrackml/pkg/common/api/response"
 	"github.com/G-Research/fasttrackml/pkg/common/middleware"
 )
