+++ pkg/common/services/artifact/storage/mock_artifact_storage_factory_provider.go
