+++ tests/integration/golang/mlflow/artifact/get_artifact_local_test.go
@@ -14,9 +14,9 @@ import (
 	"github.com/stretchr/testify/suite"
 
 	"github.com/G-Research/fasttrackml/pkg/api/mlflow"
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api/request"
 	"github.com/G-Research/fasttrackml/pkg/api/mlflow/dao/models"
 	"github.com/G-Research/fasttrackml/pkg/common/api"
+	"github.com/G-Research/fasttrackml/pkg/common/api/request"
 	"github.com/G-Research/fasttrackml/tests/integration/golang/helpers"
 )
