+++ tests/integration/golang/mlflow/artifact/list_gs_test.go
@@ -10,10 +10,10 @@ import (
 	"github.com/stretchr/testify/suite"
 
 	"github.com/G-Research/fasttrackml/pkg/api/mlflow"
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api/request"
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api/response"
 	"github.com/G-Research/fasttrackml/pkg/api/mlflow/dao/models"
 	"github.com/G-Research/fasttrackml/pkg/common/api"
+	"github.com/G-Research/fasttrackml/pkg/common/api/request"
+	"github.com/G-Research/fasttrackml/pkg/common/api/response"
 	"github.com/G-Research/fasttrackml/tests/integration/golang/helpers"
 )
